# [Heartbeat] Fix Timerqueue initialization / maintainability