--- a/heartbeat/scheduler/timerqueue/queue_test.go
+++ b/heartbeat/scheduler/timerqueue/queue_test.go
@@ -19,7 +19,6 @@ package timerqueue
 
 import (
 	"context"
-	"fmt"
 	"math/rand"
 	"sort"
 	"testing"
@@ -31,7 +30,6 @@ import (
 func TestQueueRunsInOrder(t *testing.T) {
 	// Bugs can show up only occasionally
 	for i := 0; i < 10000; i++ {
-		fmt.Printf("%d\n", i)
 		testQueueRunsInOrderOnce(t)
 	}
 }