--- a/heartbeat/tests/system/test_icmp.py
+++ b/heartbeat/tests/system/test_icmp.py
@@ -44,8 +44,7 @@ class Test(BaseTest):
         # we should run pings on those machines and make sure they work.
         self.wait_until(lambda: has_started_message() or has_failed_message(), 30)
 
-        if has_failed_message():
-            proc.check_kill_and_wait(1)
-        else:
+        if not has_failed_message():
             # Check that documents are moving through
             self.wait_until(lambda: self.output_has(lines=1))
+        proc.check_kill_and_wait()