Fixes https://github.com/elastic/beats/issues/26205

This PR changes our strategy from resetting timers to using a ticker with 250ms resolution. For whatever reason, reset seems to now be unstable (maybe a go runtime change?) on windows.

This PR also switches some hard to read parts of the code to using the `time.Until` syntax. I'm including this in this change to make the code easier to read in the future, which will help in future debugging. It also restructures some confusing code to make it easier to read.

I do suspect that there was a change in the go runtime implementation that caused this issue to service now since it's been around for a year and 1/2.

## Why is it important?

<!-- Mandatory
Explain here the WHY, or the rationale/motivation for the changes.
-->

## Checklist

<!-- Mandatory
Add a checklist of things that are required to be reviewed in order to have the PR approved

List here all the items you have verified BEFORE sending this PR. Please DO NOT remove any item, striking through those that do not apply. (Just in case, strikethrough uses two tildes. ~~Scratch this.~~)
-->

- [x] My code follows the style guidelines of this project
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have made corresponding change to the default configuration files
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I have added an entry in `CHANGELOG.next.asciidoc` or `CHANGELOG-developer.next.asciidoc`.


## How to test this PR locally

`go test -timeout 30s -run '^TestQueueRunsInOrder$' github.com/elastic/beats/v7/heartbeat/scheduler/timerqueue -count=8000` should reproduce this locally on master. It should never happen on this branch