@andrewvc 2021-06-16
Scheduler benchmarks

This PR:
```
andrewvc@LAPTOP-80O11FM2 ~/p/b/heartbeat (fix-tq)> /home/andrewvc/.gvm/versions/go1.16.4.linux.amd64/bin/go test -benchmem -bench '^(BenchmarkScheduler)$' github.com/elastic/beats/v7/heartbeat/scheduler -benchtime=20s
goos: linux
goarch: amd64
pkg: github.com/elastic/beats/v7/heartbeat/scheduler
cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
BenchmarkScheduler-10            9716433              2692 ns/op             493 B/op         14 allocs/op
PASS
ok      github.com/elastic/beats/v7/heartbeat/scheduler 28.680s
```

Master

```
andrewvc@LAPTOP-80O11FM2 ~/p/b/heartbeat (master)> /home/andrewvc/.gvm/versions/go1.16.4.linux.amd64/bin/go test -benchmem -bench '^(BenchmarkScheduler)$' github.com/elastic/beats/v7/heartbeat/scheduler -benchtime=20s
goos: linux
goarch: amd64
pkg: github.com/elastic/beats/v7/heartbeat/scheduler
cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
BenchmarkScheduler-10            7515495              2999 ns/op             504 B/op         13 allocs/op
PASS
ok      github.com/elastic/beats/v7/heartbeat/scheduler 44.524s
```